+++ chain/chain/src/stateless_validation/mod.rs
@@ -1,2 +1,5 @@
 pub(crate) mod chunk_endorsement;
+pub mod chunk_validation;
+pub mod metrics;
+pub mod processing_tracker;
 pub(crate) mod state_transition_data;
