+++ chain/client/src/stateless_validation/mod.rs
@@ -1,7 +1,6 @@
 pub mod chunk_endorsement_tracker;
 pub mod chunk_validator;
 pub mod partial_witness;
-pub mod processing_tracker;
 mod shadow_validate;
 mod state_witness_producer;
 pub mod state_witness_tracker;
